% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{anyLookBack}
\alias{anyLookBack}
\title{creates a LookBackOption class with type "any", NULL fixed lookback,
and FALSE require observe}
\usage{
anyLookBack(requiredObserved = FALSE)
}
\arguments{
\item{requiredObserved}{Require lookback time be all observed}

\item{anyLookBack}{The Analysis object to used for generation}

\item{days}{The number of days used for any look back}
}
\value{
A list containing a tibble for categorical and continuous results
}
\description{
creates a LookBackOption class with type "any", NULL fixed lookback,
and FALSE require observe
}
