% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createAnnualPrevalenceAnalysis}
\alias{createAnnualPrevalenceAnalysis}
\title{Create an AnnualPrevalenceAnalysis object}
\usage{
createAnnualPrevalenceAnalysis(
  periodOfInterest,
  reportMultiplier = 1e+05,
  populationStandardization = NULL
)
}
\arguments{
\item{periodOfInterest}{Numeric vector specifying the period of interest (required).}

\item{reportMultiplier}{Numeric value used for scaling the incidence report (optional).}

\item{populationStandardization}{A `PopulationStandardization` object (optional).}

\item{cohortOfInterest}{`CohortInfo` object of interest (required).}

\item{lookbackOptions}{A `LookBackOption` object (required).}

\item{incidenceType}{Character string: one of `"proportion"`, `"rate"`, or `"both"` (required).}

\item{strata}{A `DemographicStrata` object (optional).}
}
\value{
An `AnnualPrevalenceAnalysis` R6 object.
}
\description{
Constructs an `AnnualPrevalenceAnalysis` object with the specified settings.
}
