% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createCohortInfo}
\alias{createCohortInfo}
\title{Create a CohortInfo object}
\usage{
createCohortInfo(id, name)
}
\arguments{
\item{id}{Integer. The cohort definition ID.}

\item{name}{Character. The name of the cohort.}
}
\value{
A `CohortInfo` R6 object.
}
\description{
Constructs a `CohortInfo` object from a cohort ID and name.
}
