% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createPeriodOfInterest}
\alias{createPeriodOfInterest}
\title{Create a PeriodOfInterest object}
\usage{
createPeriodOfInterest(startYear, endYear)
}
\arguments{
\item{startYear}{Integer start year.}

\item{endYear}{Integer end year.}
}
\value{
A `PeriodOfInterest` R6 object.
}
\description{
Create a PeriodOfInterest object
}
