% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createPimsAnalysis}
\alias{createPimsAnalysis}
\title{Create a PimsAnalysis object}
\usage{
createPimsAnalysis(
  analysisCohorts,
  prevalence = NULL,
  incidence = NULL,
  mortality = NULL,
  strata = NULL
)
}
\arguments{
\item{analysisCohorts}{A list of `CohortInfo` objects (required).}

\item{prevalence}{A `PrevalenceOptions` object (optional).}

\item{incidence}{An `IncidenceOptions` object (optional).}

\item{mortality}{A `MortalityOptions` object (optional).}

\item{strata}{A `DemographicStrata` object (optional).}
}
\value{
A `PimsAnalysis` R6 object.
}
\description{
Constructs and returns a `PimsAnalysis` object
}
