% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createStrataOptions}
\alias{createStrataOptions}
\title{Create a DemographicStrata object}
\usage{
createStrataOptions(strataOptions = c("age", "gender"))
}
\arguments{
\item{strataOptions}{Character vector subset of c("age", "gender", "race", "ethnicity", "location")}
}
\value{
A `DemographicStrata` R6 object
}
\description{
Create a DemographicStrata object
}
