% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{demographicStrata}
\alias{demographicStrata}
\title{creates a demographicStrata specification}
\usage{
demographicStrata(
  strataOptions = c("age", "gender", "race", "ethnicity", "location")
)
}
\arguments{
\item{strataOptions}{Strata options: "age", "gender", "race", "ethnicity", "location"}
}
\value{
A demographic strata configuration object
}
\description{
creates a demographicStrata specification
}
