% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{fixedLookBack}
\alias{fixedLookBack}
\title{creates a LookBackOption class with type "fixed", N}
\usage{
fixedLookBack(days = 365, requiredObserved = TRUE)
}
\arguments{
\item{days}{The number of days used for fixed look back}

\item{requiredObserved}{Require lookback time be all observed}

\item{fixedLookBack}{The Analysis object to used for generation}
}
\value{
A list containing a tibble for categorical and continuous results
}
\description{
creates a LookBackOption class with type "fixed", N
}
