% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{incidence}
\alias{incidence}
\title{creates a incidence specification}
\usage{
incidence(incidenceType = c("proportion", "rate", "both"), reportMult = 1e+05)
}
\arguments{
\item{incidenceType}{incidence type: either "proportion","rate", or "both"}

\item{reportMult}{Multiplier for reporting (default: 100,000)}
}
\value{
An incidence configuration object
}
\description{
creates a incidence specification
}
