% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{mortality}
\alias{mortality}
\title{creates a mortality specification}
\usage{
mortality(mortalityType = c("earliest", "incident"), reportMult = 1e+05)
}
\arguments{
\item{mortalityType}{mortality type: either "earliest","incident"}

\item{reportMult}{Multiplier for reporting (default: 100,000)}
}
\value{
A mortality configuration object
}
\description{
creates a mortality specification
}
