% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{pointPrevalence}
\alias{pointPrevalence}
\title{creates a pointPrevalence specification}
\usage{
pointPrevalence(
  denom = c("sufficient", "complete", "Day1"),
  n = NULL,
  reportMult = 1e+05
)
}
\arguments{
\item{denom}{Denominator type: either "sufficient","Day1", or "complete"}

\item{n}{Number of days for prevalence window (required for "sufficient" denominator)}

\item{reportMult}{Multiplier for reporting (default: 100,000)}
}
\value{
A point prevalence configuration object
}
\description{
creates a pointPrevalence specification
}
