% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{pointPrevalenceDay1}
\alias{pointPrevalenceDay1}
\title{Point prevalence with "Day1" denominator}
\usage{
pointPrevalenceDay1(n = NULL, reportMult = 1e+05)
}
\description{
Point prevalence with "Day1" denominator
}
