% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{pointPrevalenceSufficient}
\alias{pointPrevalenceSufficient}
\title{Point prevalence with "sufficient" denominator}
\usage{
pointPrevalenceSufficient(n = 30, reportMult = 1e+05)
}
\description{
Point prevalence with "sufficient" denominator
}
