% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{populationStandardization}
\alias{populationStandardization}
\title{creates a population standardization specification}
\usage{
populationStandardization(weights = c("acs", "census", "jpnPop"))
}
\arguments{
\item{strataOptions}{Strata options: "acs", "census", "jpnPop"}
}
\value{
A population standardization configuration object
}
\description{
creates a population standardization specification
}
