% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{renderObservationPeriod}
\alias{renderObservationPeriod}
\title{Generate SQL to extract observation periods for a target cohort}
\usage{
renderObservationPeriod(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  startYear,
  endYear,
  required_op_length = 0,
  targetCohortId,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connectionDetails}{A `connectionDetails` object created by
`DatabaseConnector::createConnectionDetails()`}

\item{cdmDatabaseSchema}{The schema name where the data resides.}

\item{cohortDatabaseSchema}{The schema name where the cohort table is located.}

\item{cohortTable}{The name of the cohort table containing the target cohort.}

\item{startYear}{An integer specifying the start year for filtering observation periods.}

\item{endYear}{An integer specifying the end year for filtering observation periods.}

\item{required_op_length}{An integer indicating the minimum required length (in days)
of observation periods to include (default is 0, meaning no minimum length).}

\item{targetCohortId}{An integer specifying the cohort ID of interest}

\item{tempEmulationSchema}{Optional; a schema used to emulate temporary tables if the
database platform does not support them natively. Default is obtained from the
R option `"sqlRenderTempEmulationSchema"`.}
}
\description{
Renders and translates the SQL to retrieve observation period data.
}
