% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runAnnualPrevalenceAnalysis}
\alias{runAnnualPrevalenceAnalysis}
\title{Run Annual Prevalence Analysis for a Target Cohort}
\usage{
runAnnualPrevalenceAnalysis(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  targetCohort,
  prevalenceAnalysis,
  strata,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connectionDetails}{A `connectionDetails` object created by
`DatabaseConnector::createConnectionDetails()`}

\item{cdmDatabaseSchema}{The schema name where the data resides.}

\item{cohortDatabaseSchema}{The schema name where the cohort table is located.}

\item{cohortTable}{The name of the cohort table containing the target cohort.}

\item{targetCohort}{An integer specifying the cohort ID of interest.}

\item{prevalenceAnalysis}{An `AnnualPrevalenceAnalysis` object specifying incidence analysis settings.}

\item{strata}{A `DemographicStrata` object specifying the columns to stratify results by.}

\item{tempEmulationSchema}{Optional; schema used to emulate temporary tables if needed.
Defaults to the value of `getOption("sqlRenderTempEmulationSchema")`.}
}
\description{
Executes an annual prevalence analysis on a specified cohort within an OMOP CDM database.
}
