% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runPimsAnalysis}
\alias{runPimsAnalysis}
\title{Run PIMS Analysis}
\usage{
runPimsAnalysis(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  pimsAnalysis,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connectionDetails}{A `connectionDetails` object created by
`DatabaseConnector::createConnectionDetails()`}

\item{cdmDatabaseSchema}{The schema name where the data resides.}

\item{cohortDatabaseSchema}{The schema name where the cohort table is located.}

\item{cohortTable}{The name of the cohort table containing the target cohort.}

\item{pimsAnalysis}{A `PimsAnalysis` object specifying analysis settings.}

\item{tempEmulationSchema}{Optional; schema used to emulate temporary tables if needed.
Defaults to the value of `getOption("sqlRenderTempEmulationSchema")`.}
}
\description{
Executes the PIMS analysis using the specified options.
}
